+++ csatar-plugins/csatar/models/Patrol.php
@@ -91,7 +91,13 @@ class Patrol extends OrganizationBase
     public $attachOne = [
         'logo' => 'System\Models\File'
     ];
-    
+
+    public function beforeSave()
+    {
+        $filterWords = explode(',', Lang::get('csatar.csatar::lang.plugin.admin.patrol.filterOrganizationUnitNameForWords'));
+        $this->name = str_replace($filterWords, '', $this->name);
+    }
+
     /**
      * Override the getExtendedNameAttribute function
      */