+++ csatar-plugins/csatar/models/TeamReport.php
@@ -48,7 +48,7 @@ class TeamReport extends Model
     {
         // check that the team report for this team and this team doesn't already exist
         $this->year = date('n') == 1 ? date('Y') - 1 : date('Y');
-        if (TeamReport::where('team_id', $this->team_id)->where('year', $this->year)->where('deleted_at', NULL)->exists()) {
+        if (TeamReport::where('team_id', $this->team_id)->where('year', $this->year)->where('deleted_at', null)->where('id', '<>', $this->id)->exists()) {
             throw new \ValidationException(['team_id' => Lang::get('csatar.csatar::lang.plugin.component.teamReport.validationExceptions.teamReportAlreadyExists')]);
         }
 
@@ -140,7 +140,7 @@ class TeamReport extends Model
         $scouts = Scout::where('team_id', $this->team_id)->where('is_active', true)->get();
         foreach ($scouts as $scout) {
             $leadershipQualification = $scout->leadership_qualifications->sortByDesc(function ($item, $key) {
-                return $item && $item['pivot'] ? $item['pivot']['date'] : NULL;
+                return $item['pivot']['date'];
             })->values()->first();
             $membership_fee = $this->team->district->association->legal_relationships->where('id', $scout->legal_relationship_id)->first()->pivot->membership_fee;